Fix test_auth_enforced for keystonemiddleware AuthTokenPlugin removal
The keystonemiddleware change Ic4dc67662f4e1029604f23451bb75b87c1cd37c1
removed the deprecated AuthTokenPlugin fallback. Without an explicit
www_authenticate_uri, the middleware can no longer build the
WWW-Authenticate header for 401 responses, causing a 500 error instead.

Add www_authenticate_uri to the test paste config authtoken filter.

Fixed pep8 error as this change is blocking all CI.

Closes-Bug: #2153160

+++ neutron/conf/agent/database/agents_db.py
@@ -21,7 +21,7 @@ AGENT_OPTS = [
     # eventlet in milliseconds and any number higher will produce an OverFlow
     # error. More details here: https://bugs.launchpad.net/neutron/+bug/2028724
     cfg.IntOpt('agent_down_time', default=75,
-               max=((2**32 / 2 - 1) // 1000),
+               max=int((2**32 / 2 - 1) // 1000),
                help=_("Seconds to regard the agent as down; should be at "
                       "least twice report_interval, to be sure the "
                       "agent is down for good.")),
